Refactor QRLoader container
