+++ src/components/BcmcMobile/config.ts
