+++ src/components/Swish/Swish.ts
@@ -1,8 +1,16 @@
-import withQRLoader from '../helpers/withQRLoader';
+import QRLoaderContainer from '../helpers/QRLoaderContainer';
 
-export default withQRLoader({
-    type: 'swish',
-    shouldRedirectOnMobile: true,
-    STATUS_INTERVAL: 2000, // ms
-    COUNTDOWN_MINUTES: 3 // min
-});
+class SwishElement extends QRLoaderContainer {
+    public static type = 'swish';
+
+    formatProps(props) {
+        return {
+            shouldRedirectOnMobile: true,
+            STATUS_INTERVAL: 2000, // ms
+            COUNTDOWN_MINUTES: 3, // min
+            ...super.formatProps(props)
+        };
+    }
+}
+
+export default SwishElement;
