--- src/components/WeChat/WeChat.js
@@ -1,7 +1,0 @@
-import withQRLoader from '../helpers/withQRLoader';
-import * as config from './config';
-
-export default withQRLoader({
-    type: 'wechatpayQR',
-    ...config
-});
