+++ src/components/WeChat/WeChat.test.ts
