+++ src/components/WeChat/WeChat.ts
@@ -1,0 +1,16 @@
+import QRLoaderContainer from '../helpers/QRLoaderContainer';
+import { delay, countdownTime } from './config';
+
+class WeChatPayElement extends QRLoaderContainer {
+    public static type = 'wechatpayQR';
+
+    formatProps(props) {
+        return {
+            delay,
+            countdownTime,
+            ...super.formatProps(props)
+        };
+    }
+}
+
+export default WeChatPayElement;
