+++ src/components/WeChat/index.ts
